# gen/http/openapi3.json
{"openapi":"3.0.3","info":{"title":"XFSC Status List Service","description":"Status list REST API","version":"0.0.1"},"servers":[{"url":"http://localhost:8080"}],"paths":{"/health":{"get":{"tags":["status"],"summary":"health status","operationId":"status#health","responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"type":"string","example":"Necessitatibus doloribus in veritatis qui."},"example":"Ipsa voluptates distinctio nihil eos dolor."}}}}}},"/status/{listId}":{"get":{"tags":["status"],"summary":"getList status","description":"Returns a status list as JSON, JWT status list, or StatusList2021 credential depending on Content-Type.","operationId":"status#getList","parameters":[{"name":"listId","in":"path","required":true,"schema":{"type":"integer","example":2113381193853987314,"format":"int64"},"example":793274971772924104},{"name":"X-Tenant-Id","in":"header","allowEmptyValue":true,"required":true,"schema":{"type":"string","example":"Soluta tempora omnis iure et aut."},"example":"Sit non ut nostrum et qui impedit."},{"name":"Accept","in":"header","allowEmptyValue":true,"schema":{"type":"string","example":"Aliquid in non ut et magni."},"example":"Sit eaque dolore commodi temporibus qui sunt."}],"responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"example":"Est sunt fuga a."},"example":"Dolores et aspernatur perferendis quasi reiciendis."}}}}}},"/status/{tenantId}/{listId}/revoke/{index}":{"post":{"tags":["status"],"summary":"revoke status","description":"Revokes a credential by setting the bit at the given index.","operationId":"status#revoke","parameters":[{"name":"tenantId","in":"path","required":true,"schema":{"type":"string","example":"Officia et repudiandae et molestiae labore."},"example":"Tempore ad voluptas nihil cumque corporis."},{"name":"listId","in":"path","required":true,"schema":{"type":"integer","example":2610346682166145381,"format":"int64"},"example":9180807483828302258},{"name":"index","in":"path","required":true,"schema":{"type":"integer","example":7861203698842419081,"format":"int64"},"example":969487697790967545}],"responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/XfscStatusListRevoke"},"example":{"index":3930967623290764634,"listId":5730539016118596574,"status":"Aut a nesciunt.","tenantId":"Et laboriosam vel sunt magni id voluptatum."}}}},"400":{"description":"bad_request: Bad Request response.","content":{"application/vnd.xfsc.error":{"schema":{"$ref":"#/components/schemas/XfscError"},"example":{"message":"Qui accusamus.","status":1886340408761926525}}}},"500":{"description":"internal_error: Internal Server Error response.","content":{"application/vnd.xfsc.error":{"schema":{"$ref":"#/components/schemas/XfscError"},"example":{"message":"Dolor accusantium ab officia odit quis.","status":4629432125991665016}}}}}}}},"components":{"schemas":{"XfscError":{"type":"object","properties":{"message":{"type":"string","example":"Qui adipisci sed est."},"status":{"type":"integer","example":6184804881049467816,"format":"int64"}},"example":{"message":"Pariatur saepe voluptatem.","status":570516531640558798},"required":["message","status"]},"XfscStatusList":{"type":"object","properties":{"list":{"type":"string","example":"Illo sit natus."},"listId":{"type":"integer","example":8068686243924224578,"format":"int64"},"tenantId":{"type":"string","example":"Nobis quod velit placeat animi tempore."}},"example":{"list":"Vel ex aut ratione et sit et.","listId":1663714144915302481,"tenantId":"Sed rem."},"required":["tenantId","listId","list"]},"XfscStatusListRevoke":{"type":"object","properties":{"index":{"type":"integer","example":3421304832824289482,"format":"int64"},"listId":{"type":"integer","example":8432596197339287520,"format":"int64"},"status":{"type":"string","example":"Tenetur aut assumenda."},"tenantId":{"type":"string","example":"Accusamus nihil ut nobis."}},"example":{"index":290659539145452584,"listId":4248965308005080749,"status":"Rem dignissimos et.","tenantId":"Aut aut a voluptates."},"required":["tenantId","listId","index","status"]}}},"tags":[{"name":"status","description":"Status list endpoints"}]}