# gen/http/status/server/encode_decode.go
// Code generated by goa, DO NOT EDIT.
//
// status HTTP server encoders and decoders
//
// Command:
// $ goa gen github.com/eclipse-xfsc/statuslist-service/design

package server

import (
	"context"
	"errors"
	"net/http"
	"strconv"

	status "github.com/eclipse-xfsc/statuslist-service/gen/status"
	statusviews "github.com/eclipse-xfsc/statuslist-service/gen/status/views"
	goahttp "goa.design/goa/v3/http"
	goa "goa.design/goa/v3/pkg"
)

// EncodeHealthResponse returns an encoder for responses returned by the status
// health endpoint.
func EncodeHealthResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.(string)
		enc := encoder(ctx, w)
		body := res
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// EncodeGetListResponse returns an encoder for responses returned by the
// status getList endpoint.
func EncodeGetListResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.(any)
		enc := encoder(ctx, w)
		body := res
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// DecodeGetListRequest returns a decoder for requests sent to the status
// getList endpoint.
func DecodeGetListRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (*status.GetListPayload, error) {
	return func(r *http.Request) (*status.GetListPayload, error) {
		var payload *status.GetListPayload
		var (
			listID   int
			tenantID string
			accept   *string
			err      error

			params = mux.Vars(r)
		)
		{
			listIDRaw := params["listId"]
			v, err2 := strconv.ParseInt(listIDRaw, 10, strconv.IntSize)
			if err2 != nil {
				err = goa.MergeErrors(err, goa.InvalidFieldTypeError("listId", listIDRaw, "integer"))
			}
			listID = int(v)
		}
		tenantID = r.Header.Get("X-Tenant-Id")
		if tenantID == "" {
			err = goa.MergeErrors(err, goa.MissingFieldError("tenantId", "header"))
		}
		acceptRaw := r.Header.Get("Accept")
		if acceptRaw != "" {
			accept = &acceptRaw
		}
		if err != nil {
			return payload, err
		}
		payload = NewGetListPayload(listID, tenantID, accept)

		return payload, nil
	}
}

// EncodeRevokeResponse returns an encoder for responses returned by the status
// revoke endpoint.
func EncodeRevokeResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res := v.(*statusviews.RevokeResult)
		enc := encoder(ctx, w)
		body := NewRevokeResponseBody(res.Projected)
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// DecodeRevokeRequest returns a decoder for requests sent to the status revoke
// endpoint.
func DecodeRevokeRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (*status.RevokePayload, error) {
	return func(r *http.Request) (*status.RevokePayload, error) {
		var payload *status.RevokePayload
		var (
			tenantID string
			listID   int
			index    int
			err      error

			params = mux.Vars(r)
		)
		tenantID = params["tenantId"]
		{
			listIDRaw := params["listId"]
			v, err2 := strconv.ParseInt(listIDRaw, 10, strconv.IntSize)
			if err2 != nil {
				err = goa.MergeErrors(err, goa.InvalidFieldTypeError("listId", listIDRaw, "integer"))
			}
			listID = int(v)
		}
		{
			indexRaw := params["index"]
			v, err2 := strconv.ParseInt(indexRaw, 10, strconv.IntSize)
			if err2 != nil {
				err = goa.MergeErrors(err, goa.InvalidFieldTypeError("index", indexRaw, "integer"))
			}
			index = int(v)
		}
		if err != nil {
			return payload, err
		}
		payload = NewRevokePayload(tenantID, listID, index)

		return payload, nil
	}
}

// EncodeRevokeError returns an encoder for errors returned by the revoke
// status endpoint.
func EncodeRevokeError(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder, formatter func(ctx context.Context, err error) goahttp.Statuser) func(context.Context, http.ResponseWriter, error) error {
	encodeError := goahttp.ErrorEncoder(encoder, formatter)
	return func(ctx context.Context, w http.ResponseWriter, v error) error {
		var en goa.GoaErrorNamer
		if !errors.As(v, &en) {
			return encodeError(ctx, w, v)
		}
		switch en.GoaErrorName() {
		case "bad_request":
			var res *status.ErrorResult
			errors.As(v, &res)
			ctx = context.WithValue(ctx, goahttp.ContentTypeKey, "application/vnd.xfsc.error")
			enc := encoder(ctx, w)
			var body any
			if formatter != nil {
				body = formatter(ctx, res)
			} else {
				body = NewRevokeBadRequestResponseBody(res)
			}
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusBadRequest)
			return enc.Encode(body)
		case "internal_error":
			var res *status.ErrorResult
			errors.As(v, &res)
			ctx = context.WithValue(ctx, goahttp.ContentTypeKey, "application/vnd.xfsc.error")
			enc := encoder(ctx, w)
			var body any
			if formatter != nil {
				body = formatter(ctx, res)
			} else {
				body = NewRevokeInternalErrorResponseBody(res)
			}
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusInternalServerError)
			return enc.Encode(body)
		default:
			return encodeError(ctx, w, v)
		}
	}
}